+++ packages/core/CHANGELOG.md
@@ -1,5 +1,14 @@
 # web-map-service
 
+## 0.2.2
+
+### Patch Changes
+
+- 发布测试
+- Updated dependencies
+  - @web-map-service/map2d-app@0.2.2
+  - @web-map-service/map2d@0.2.2
+
 ## 0.2.1
 
 ### Patch Changes