+++ packages/map2d/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @web-map-service/map2d
 
+## 0.2.2
+
+### Patch Changes
+
+- 发布测试
+
 ## 0.2.1
 
 ### Patch Changes